refactored and broken down components
