+++ src/pages/home/home.css
@@ -1,25 +1,44 @@
 .hero{
     margin-top: 5rem;
+    /* width: 100%; */
     /* width: 700px;     */
     
     /* border:1px solid red; */
 }
 .hero button{
-    font-size: 2rem; 
+    font-size: 1.5rem; 
     background-color:#9CA3AF;
     font-family: inherit;
     color: var(--white);
     border:none;
     border-radius: 4px;
-    margin:1rem;
-    padding: 1rem;
+    margin: 0.5em;
+        padding: 0.5rem;
 }
 .hero button:hover{
     cursor: pointer;
+}
+.hero-content{
+
 }
 .hero p{
     font-size: 2rem;
     margin:1rem;
     padding:1rem;
     color : #9CA3AF;
+   
+    display: block;
+    text-align: center;
+
+}
+.hero img{
+    width: 100%;
+    height:50vh;
+    
+}
+.image-container{
+    width: 100%;
+   
+    
+    
 }
