+++ src/pages/products-listing/products-listing.css
@@ -1,3 +1,5 @@
+
+
 .product-listing,
 .cart-listing {
   /* border: 1px solid red; */
@@ -5,6 +7,17 @@
   display: grid;
   grid-template-columns: 1fr 1fr 1fr 1fr;
   /* order:2; */
+}
+.list-filter-button{
+  display: flex;
+  flex-direction: column;
+  position: relative;
+}
+.filter-button{
+  position: fixed;
+  top:5%;
+  left:40%;
+  border:1px solid red;
 }
 
 /**card*/
@@ -100,6 +113,9 @@ aside label{
   display: grid;
   grid-template-columns: 1fr 1fr ;
   /* order:2; */
+}
+.filter-button{
+  display: absolute;
 }
     
 }
