+++ src/pages/products-listing/products-listing.jsx
@@ -1,5 +1,7 @@
 import './products-listing.css';
 //import data from '../../data';
+import { Card } from '../../components/Card/Card';
+import { Filter } from '../../components/Filters/Filter'
 
 import { useCart } from '../../context/cart-context';
 import { useProduct } from '../../context/product-context';
@@ -69,97 +71,20 @@ export function ProductListing() {
     return (
         <>
             <div className="product-section">
-                <div className="div-sidebar">
-                    <aside>
-                        <p className="filter-heading">Sort By</p>
-                        <label>
-                            <input
-                                type="radio"
-                                name="sort"
-                                onChange={() => productFilterDispatch({ type: "SORT", payload: "HIGH_TO_LOW" })}
-                                checked={sortBy === "HIGH_TO_LOW"}
-                            />
-                            Price High to Low
-                        </label>
-                        <label>
-                            <input
-                                type="radio"
-                                name="sort"
-                                onChange={() => productFilterDispatch({ type: "SORT", payload: "LOW_TO_HIGH" })}
-                                checked={sortBy === "LOW_TO_HIGH"}
-                            />
-                            Price Low to High
-                        </label>
-                        <p className="filter-heading">Filters</p>
-
-                        <label>
-                            <input
-                                type="checkbox"
-                                checked={showAllInventory}
-                                onChange={() => productFilterDispatch({ type: "TOGGLE_INVENTORY" })}
-                            />
-            Include Out of Stock
-          </label>
-
-                        <label>
-                            <input
-                                type="checkbox"
-                                checked={showFastDelivery}
-                                onChange={() => productFilterDispatch({ type: "TOGGLE_DELIVERY" })}
-                            />
-                            Fast Delivery Only
-          </label>
-                        {/* <button onClick={() => productFilterDispatch({ type: "CLEAR_FILTERS" })}>Clear All Filters</button> */}
-                    </aside>
+                <Filter sortBy={sortBy} showAllInventory={showAllInventory} showFastDelivery={showFastDelivery} productFilterDispatch={productFilterDispatch} />
+                <div className="list-filter-button">
+                    <div className="product-listing">
+                        {filteredData.map((product) => (
+                            <Card key={product.id} product={product} getColor={getColor} isItemInWishList={isItemInWishList} isItemInCart={isItemInCart}
+                                cartList={cartList} wishList={wishList} dispatch={dispatch} />
+                        ))}
+                    </div>
+                    {/* <div className="filter-button">
+                        <button>Filter</button>
+                        <button>Sort</button>
+                    </div> */}
                 </div>
-                <div className="product-listing">
-                    {filteredData.map((product) => (
-                        <div className="card card-badge" key={product.id}>
-                            <img src={product.image} alt="pic" />
 
-                            <div className="card-content">
-                                <div className="div-brand-name">
-                                    <p className="para-price">{product.name}</p>
-                                    <span
-                                        className="material-icons icons"
-                                        style={{
-                                            color: getColor(wishList, product.id),
-                                            border: "black"
-                                        }}
-                                        onClick={() =>
-                                            isItemInWishList(wishList, product.id)
-                                                ? dispatch({
-                                                    type: "REMOVE_FROM_WISHLIST",
-                                                    payload: product
-                                                })
-                                                : dispatch({ type: "ADD_TO_WISHLIST", payload: product })
-                                        }
-                                    >
-                                        favorite
-                                    </span>
-                                </div>
-                                <p className="card-text">{product.material}</p>
-                                <h4 className="heading price"><span>&#8377;{product.price}</span></h4>
-
-                                {isItemInCart(cartList, product.id) ? (
-                                    <button className="btn btn-secondary btn-disabled" disabled>
-                                        <span>Added</span>
-                                        <span class="material-icons added-icon ">check_circle</span>
-                                    </button>
-                                ) : (
-                                    <button
-                                        className="btn btn-primary"
-                                        onClick={() =>
-                                            dispatch({ type: "ADD_TO_CART", payload: product })
-                                        }
-                                    >
-                                        Add To Cart
-                                    </button>
-                                )}
-                            </div>
-                        </div>
-                    ))}
-                </div>
             </div>
         </>
     )
